+++ src/demo/demo/config-form/config-form.component.ts
@@ -25,7 +25,8 @@ const PICKER_OPTION_KEYS = [
   'placeholder',
   'required',
   'hideInputContainer',
-  'hideOnOutsideClick'
+  'hideOnOutsideClick',
+  'closeOnEnter'
 ];
 const DAY_PICKER_DIRECTIVE_OPTION_KEYS = [
   'allowMultiSelect',
@@ -58,6 +59,7 @@ const MONTH_CALENDAR_OPTION_KEYS = [
   'showGoToCurrent',
   'unSelectOnClick',
   'moveCalendarTo',
+  'numOfMonthRows',
   ...GLOBAL_OPTION_KEYS
 ];
 const DAY_CALENDAR_OPTION_KEYS = [
@@ -219,6 +221,8 @@ export class ConfigFormComponent implements OnInit {
   showMultipleYearsNavigation: FormControl;
   multipleYearsNavigateBy: FormControl;
   returnedValueType: FormControl;
+  closeOnEnter: FormControl;
+  numOfMonthRows: FormControl;
 
   ngOnInit() {
     this.localFormat = this.getDefaultFormatByMode(this.pickerMode);
@@ -264,6 +268,8 @@ export class ConfigFormComponent implements OnInit {
     this.showMultipleYearsNavigation = new FormControl(this.config.showMultipleYearsNavigation);
     this.multipleYearsNavigateBy = new FormControl(this.config.multipleYearsNavigateBy);
     this.returnedValueType = new FormControl(this.config.returnedValueType);
+    this.closeOnEnter = new FormControl(this.config.closeOnEnter);
+    this.numOfMonthRows = new FormControl(this.config.numOfMonthRows);
     this.initListeners();
   }
 
@@ -607,6 +613,18 @@ export class ConfigFormComponent implements OnInit {
     this.returnedValueType.valueChanges.subscribe((val) => {
       this.onConfigChange.emit({
         returnedValueType: val
+      });
+    });
+
+    this.closeOnEnter.valueChanges.subscribe((val) => {
+      this.onConfigChange.emit({
+        closeOnEnter: val
+      });
+    });
+
+    this.numOfMonthRows.valueChanges.subscribe((val) => {
+      this.onConfigChange.emit({
+        numOfMonthRows: val
       });
     });
   }