+++ src/lib/date-picker/date-picker.service.ts
@@ -17,6 +17,7 @@ export class DatePickerService {
   private defaultConfig: IDatePickerConfigInternal = {
     closeOnSelect: true,
     closeOnSelectDelay: 100,
+    closeOnEnter: true,
     format: 'DD-MM-YYYY',
     openOnFocus: true,
     openOnClick: true,
@@ -82,7 +83,8 @@ export class DatePickerService {
       locale: pickerConfig.locale,
       returnedValueType: pickerConfig.returnedValueType,
       showGoToCurrent: pickerConfig.showGoToCurrent,
-      unSelectOnClick: pickerConfig.unSelectOnClick
+      unSelectOnClick: pickerConfig.unSelectOnClick,
+      numOfMonthRows: pickerConfig.numOfMonthRows
     };
   }